+++ src/hwenc_v4l2/v4l2_native_buffer.h
@@ -1,7 +1,11 @@
 #ifndef V4L2_NATIVE_BUFFER_H_
 #define V4L2_NATIVE_BUFFER_H_
 
+#include <functional>
+
 // WebRTC
+#include <api/scoped_refptr.h>
+#include <api/video/video_frame_buffer.h>
 #include <common_video/include/video_frame_buffer.h>
 #include <common_video/libyuv/include/webrtc_libyuv.h>
 
@@ -16,24 +20,7 @@ class V4L2NativeBuffer : public webrtc::VideoFrameBuffer {
                    const uint8_t* data,
                    int size,
                    int stride,
-                   std::function<void()> on_destruction)
-      : video_type_(video_type),
-        raw_width_(raw_width),
-        raw_height_(raw_height),
-        scaled_width_(scaled_width),
-        scaled_height_(scaled_height),
-        fd_(fd),
-        size_(size),
-        stride_(stride) {
-    if (data != nullptr) {
-      data_.reset(new uint8_t[size_]);
-      memcpy(data_.get(), data, size_);
-    }
-    if (on_destruction != nullptr) {
-      shared_on_destruction_.reset(
-          new int(), [on_destruction](int*) { on_destruction(); });
-    }
-  }
+                   std::function<void()> on_destruction);
 
   V4L2NativeBuffer(webrtc::VideoType video_type,
                    int raw_width,
@@ -44,48 +31,29 @@ class V4L2NativeBuffer : public webrtc::VideoFrameBuffer {
                    const std::shared_ptr<uint8_t> data,
                    int size,
                    int stride,
-                   std::shared_ptr<void> shared_on_destruction)
-      : video_type_(video_type),
-        raw_width_(raw_width),
-        raw_height_(raw_height),
-        scaled_width_(scaled_width),
-        scaled_height_(scaled_height),
-        fd_(fd),
-        data_(data),
-        size_(size),
-        stride_(stride),
-        shared_on_destruction_(shared_on_destruction) {}
+                   std::shared_ptr<void> shared_on_destruction);
 
-  webrtc::VideoFrameBuffer::Type type() const override {
-    return webrtc::VideoFrameBuffer::Type::kNative;
-  };
-  int width() const override { return scaled_width_; }
-  int height() const override { return scaled_height_; }
-  rtc::scoped_refptr<webrtc::I420BufferInterface> ToI420() override {
-    RTC_LOG(LS_ERROR) << "V4L2NativeBuffer::ToI420() not implemented";
-    return nullptr;
-  }
+  webrtc::VideoFrameBuffer::Type type() const override;
+  int width() const override;
+  int height() const override;
+  rtc::scoped_refptr<webrtc::I420BufferInterface> ToI420() override;
 
   // crop は無視してサイズだけ変更する
-  rtc::scoped_refptr<VideoFrameBuffer> CropAndScale(
+  rtc::scoped_refptr<webrtc::VideoFrameBuffer> CropAndScale(
       int offset_x,
       int offset_y,
       int crop_width,
       int crop_height,
       int scaled_width,
-      int scaled_height) override {
-    return rtc::make_ref_counted<V4L2NativeBuffer>(
-        video_type_, raw_width_, raw_height_, scaled_width, scaled_height, fd_,
-        data_, size_, stride_, shared_on_destruction_);
-  }
+      int scaled_height) override;
 
-  webrtc::VideoType video_type() const { return video_type_; }
-  int fd() const { return fd_; }
-  std::shared_ptr<uint8_t> data() const { return data_; }
-  int size() const { return size_; }
-  int stride() const { return stride_; }
-  int raw_width() const { return raw_width_; }
-  int raw_height() const { return raw_height_; }
+  webrtc::VideoType video_type() const;
+  int fd() const;
+  std::shared_ptr<uint8_t> data() const;
+  int size() const;
+  int stride() const;
+  int raw_width() const;
+  int raw_height() const;
 
  private:
   webrtc::VideoType video_type_;
